Implemented cyclical_serialization_perf.py
