+++ bindings/py/src/nupic/bindings/tools/SerializationTestPyRegionProto.capnp
