+++ bindings/py/tests/network_test.py
@@ -29,139 +29,12 @@ try:
 except ImportError:
   capnp = None
 else:
-  from SerializationTestPyRegionProto_capnp import SerializationTestPyRegionProto
   from nupic.proto.NetworkProto_capnp import NetworkProto
 
 
 import nupic.bindings.engine_internal as engine
-from nupic.bindings.math import Random
-from nupic.bindings.regions.PyRegion import PyRegion
-
-
-
-class SerializationTestPyRegion(PyRegion):
-  """Custom python region for testing serialization/deserialization of network
-  containing a python region.
-  """
-  def __init__(self, dataWidth, randomSeed):
-    if dataWidth <= 0:
-      raise ValueError("Parameter dataWidth must be > 0")
-
-    # Arbitrary value that's compatible with UInt32 in the proto schema
-    # for testing serialization of python-native property
-    self._dataWidth = dataWidth
-
-    # For testing serialization of object implemented in the extension
-    self._rand = Random(randomSeed)
-
-
-  @property
-  def dataWidth(self):
-    return self._dataWidth
-
-
-  @property
-  def randomSeed(self):
-    return self._rand.getSeed()
-
-
-  @staticmethod
-  def getProtoType():
-    """Return the pycapnp proto type that the class uses for serialization.
-
-    This is used to convert the proto into the proper type before passing it
-    into the read or write method of the subclass.
-    """
-    return SerializationTestPyRegionProto
-
-
-  def writeToProto(self, proto):
-    """Write state to proto object.
-
-    The type of proto is determined by getProtoType().
-    """
-    proto.dataWidth = self._dataWidth
-    self._rand.write(proto.random)
-
-
-  @classmethod
-  def readFromProto(cls, proto):
-    """Read state from proto object.
-
-    The type of proto is determined by getProtoType().
-
-    :returns: Instance of SerializationTestPyRegion initialized from proto
-    """
-    obj = object.__new__(cls)
-    obj._dataWidth = proto.dataWidth
-    obj._rand = Random()
-    obj._rand.read(proto.random)
-
-    return obj
-
-
-  def initialize(self, dims=None, splitterMaps=None):
-    pass
-
-
-  def compute(self, inputs, outputs):
-    """
-    Run one iteration of SerializationTestPyRegion's compute
-    """
-    outputs["out"][:] = inputs["in"]
-
-
-  @classmethod
-  def getSpec(cls):
-    """Return the Spec for SerializationTestPyRegion.
-    """
-    spec = {
-        "description":SerializationTestPyRegion.__doc__,
-        "singleNodeOnly":True,
-        "inputs":{
-          "in":{
-            "description":"The input vector.",
-            "dataType":"Real32",
-            "count":0,
-            "required":True,
-            "regionLevel":False,
-            "isDefaultInput":True,
-            "requireSplitterMap":False},
-        },
-        "outputs":{
-          "out":{
-            "description":"A copy of the input vector.",
-            "dataType":"Real32",
-            "count":0,
-            "regionLevel":True,
-            "isDefaultOutput":True},
-        },
-
-        "parameters":{
-          "dataWidth":{
-            "description":"Size of inputs",
-            "accessMode":"Read",
-            "dataType":"UInt32",
-            "count":1,
-            "constraints":""},
-          "randomSeed":{
-            "description":"Seed for constructing the Random instance",
-            "accessMode":"Read",
-            "dataType":"UInt32",
-            "count":1,
-            "constraints":""},
-        },
-
-    }
-
-    return spec
-
-
-  def getOutputElementCount(self, name):
-    if name == "out":
-      return self._dataWidth
-    else:
-      raise Exception("Unrecognized output: " + name)
+from nupic.bindings.tools.serialization_test_py_region import \
+     SerializationTestPyRegion
 
 
 
@@ -177,11 +50,11 @@ class NetworkTest(unittest.TestCase):
     try:
       srcNet = engine.Network()
       srcNet.addRegion(SerializationTestPyRegion.__name__,
-                        "py." + SerializationTestPyRegion.__name__,
-                        json.dumps({
-                            "dataWidth": 128,
-                            "randomSeed": 99,
-                          }))
+                       "py." + SerializationTestPyRegion.__name__,
+                       json.dumps({
+                         "dataWidth": 128,
+                         "randomSeed": 99,
+                       }))
 
       # Serialize
       builderProto = NetworkProto.new_message()
